src/main/java/SellingItems/SheetMusic.java
package SellingItems;

import Behaviours.ISell;

public class SheetMusic implements ISell {

    private String description;
    private double buyPrice;
    private double sellPrice;

    public SheetMusic(String description, double buyPrice, double sellPrice) {
        this.description = description;
        this.buyPrice = buyPrice;
        this.sellPrice = sellPrice;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public double getBuyPrice() {
        return buyPrice;
    }

    public void setBuyPrice(double buyPrice) {
        this.buyPrice = buyPrice;
    }

    public double getSellPrice() {
        return sellPrice;
    }

    public void setSellPrice(double sellPrice) {
        this.sellPrice = sellPrice;
    }

    public double calculateMarkup() {
        return this.getSellPrice() - this.getBuyPrice() ;
    }
}
